Вынесена в отдельный компонент информация о фильме.
Чтобы при переключении табов не перерисовывалась вся страница.
